Start changing trait, and method calls
