+++ crates/batch-prover/src/da_block_handler.rs
@@ -44,7 +44,7 @@ where
     Da: DaService,
     Vm: ZkvmHost + Zkvm,
     DB: BatchProverLedgerOps,
-    Ps: ProverService<Vm>,
+    Ps: ProverService,
     StateRoot: BorshDeserialize
         + BorshSerialize
         + Serialize
@@ -72,7 +72,7 @@ impl<Vm, Da, Ps, DB, StateRoot, Witness> L1BlockHandler<Vm, Da, Ps, DB, StateRoo
 where
     Da: DaService,
     Vm: ZkvmHost + Zkvm,
-    Ps: ProverService<Vm, DaService = Da>,
+    Ps: ProverService<DaService = Da>,
     DB: BatchProverLedgerOps + Clone + 'static,
     StateRoot: BorshDeserialize
         + BorshSerialize
@@ -237,7 +237,7 @@ where
 
             if should_prove {
                 if l1_height >= self.skip_submission_until_l1 {
-                    prove_l1(
+                    prove_l1::<Da, Ps, Vm, DB, StateRoot, Witness>(
                         self.da_service.clone(),
                         self.prover_service.clone(),
                         self.ledger_db.clone(),
@@ -269,9 +269,7 @@ where
 
     async fn check_and_recover_ongoing_proving_sessions(&self) -> Result<(), anyhow::Error> {
         let prover_service = self.prover_service.as_ref();
-        let results = prover_service
-            .recover_proving_sessions_and_send_to_da(&self.da_service)
-            .await?;
+        let results = prover_service.recover_and_submit_proving_sessions().await?;
 
         for (tx_id, proof) in results {
             extract_and_store_proof::<DB, Da, Vm, StateRoot>(
