--- crates/prover-services/src/parallel/new.rs
@@ -1,252 +1,0 @@
-use std::ops::DerefMut;
-use std::sync::Arc;
-
-use async_trait::async_trait;
-use borsh::BorshDeserialize;
-use futures::future;
-use parking_lot::Mutex;
-use sov_db::ledger_db::LedgerDB;
-use sov_rollup_interface::da::{DaData, DaSpec};
-use sov_rollup_interface::services::da::DaService;
-use sov_rollup_interface::stf::StateTransitionFunction;
-use sov_rollup_interface::zk::{Proof, ZkvmHost};
-use sov_stf_runner::{
-    ProofProcessingStatus, ProverService, ProverService2, ProverServiceError, WitnessSubmissionStatus
-};
-use tokio::sync::oneshot;
-
-use crate::ProofGenMode;
-
-pub(crate) type Input = Vec<u8>;
-pub(crate) type Assumptions = Vec<Vec<u8>>;
-pub(crate) type ProofData = (Input, Assumptions);
-
-/// Prover service that generates proofs in parallel.
-pub struct ParallelProverService<Da, Vm, Stf>
-where
-    Da: DaService,
-    Vm: ZkvmHost + 'static,
-    Stf: StateTransitionFunction<Vm::Guest, Da::Spec> + Send + Sync + 'static,
-    Stf::PreState: Clone + Send + Sync + 'static,
-{
-    thread_pool: rayon::ThreadPool,
-
-    proof_mode: Arc<Mutex<ProofGenMode<Da, Vm, Stf>>>,
-
-    da_service: Da,
-    vm: Vm,
-    zk_storage: Stf::PreState,
-    ledger_db: LedgerDB,
-
-    proof_queue: Vec<ProofData>,
-}
-
-impl<Da, Vm, Stf> ParallelProverService<Da, Vm, Stf>
-where
-    Da: DaService,
-    Vm: ZkvmHost,
-    Stf: StateTransitionFunction<Vm::Guest, Da::Spec> + Send + Sync,
-    Stf::PreState: Clone + Send + Sync,
-{
-    /// Creates a new prover.
-    pub fn new(
-        da_service: Da,
-        vm: Vm,
-        proof_mode: ProofGenMode<Da, Vm, Stf>,
-        zk_storage: Stf::PreState,
-        thread_pool_size: usize,
-        ledger_db: LedgerDB,
-    ) -> anyhow::Result<Self> {
-        assert!(
-            thread_pool_size > 0,
-            "Prover thread pool size must be greater than 1"
-        );
-
-        match proof_mode {
-            ProofGenMode::Skip => {
-                tracing::info!("Prover is configured to skip proving");
-            }
-            ProofGenMode::Simulate(_) => {
-                tracing::info!("Prover is configured to simulate proving");
-            }
-            ProofGenMode::Execute => {
-                tracing::info!("Prover is configured to execute proving");
-            }
-            ProofGenMode::Prove => {
-                tracing::info!("Prover is configured to prove");
-            }
-        };
-
-        let thread_pool = rayon::ThreadPoolBuilder::new()
-            .num_threads(thread_pool_size)
-            .build()
-            .expect("Thread pool must be built");
-
-        Ok(Self {
-            thread_pool,
-            proof_mode: Arc::new(Mutex::new(proof_mode)),
-            da_service,
-            vm,
-            zk_storage,
-            ledger_db,
-            proof_queue: vec![],
-        })
-    }
-
-    /// Creates a new `ParallelProverService` with thread_pool_size retrieved from
-    /// environment variable `PARALLEL_PROOF_LIMIT`. If non-existent, will panic.
-    pub fn new_from_env(
-        da_service: Da,
-        vm: Vm,
-        proof_mode: ProofGenMode<Da, Vm, Stf>,
-        zk_storage: Stf::PreState,
-        ledger_db: LedgerDB,
-    ) -> anyhow::Result<Self> {
-        let thread_pool_size = std::env::var("PARALLEL_PROOF_LIMIT")
-            .expect("PARALLEL_PROOF_LIMIT must be set")
-            .parse::<usize>()
-            .expect("PARALLEL_PROOF_LIMIT must be valid unsigned number");
-
-        Self::new(
-            da_service,
-            vm,
-            proof_mode,
-            zk_storage,
-            thread_pool_size,
-            ledger_db,
-        )
-    }
-
-
-    async fn prove(&self, proof_queue: Vec<ProofData>) -> Vec<Proof> {
-        let num_threads = self.thread_pool.current_num_threads();
-
-        // Future buffer to keep track of ongoing provings
-        let mut ongoing_proofs = Vec::with_capacity(num_threads);
-        let mut proofs = vec![Proof::default(); proof_queue.len()];
-        // Initialize proof workers
-        for (idx, proof_data) in proof_queue.into_iter().enumerate() {
-            if ongoing_proofs.len() == num_threads {
-                // If no available threads, wait for one of the proofs to finish
-                let ((idx, proof), _, remaining_proofs) = future::select_all(ongoing_proofs).await;
-                proofs[idx] = proof;
-                ongoing_proofs = remaining_proofs;
-            }
-
-            let proof_fut = self.prove_with_data(proof_data);
-            ongoing_proofs.push(Box::pin(async move {
-                let proof = proof_fut.await;
-                (idx, proof)
-            }));
-        }
-
-        // Wait for all the remaining proofs to complete
-        let remaining_proofs = future::join_all(ongoing_proofs).await;
-        for (idx, proof) in remaining_proofs {
-            proofs[idx] = proof;
-        }
-
-        proofs
-    }
-
-    async fn prove_with_data(&self, (input, assumptions): ProofData) -> Proof {
-        let mut vm = self.vm.clone();
-        let zk_storage = self.zk_storage.clone();
-        let proof_mode = self.proof_mode.clone();
-
-        vm.add_hint(input);
-        for assumption in assumptions {
-            vm.add_assumption(assumption);
-        }
-
-        let (tx, rx) = oneshot::channel();
-        self.thread_pool.spawn(move || {
-            let proof =
-                make_proof(vm, zk_storage, proof_mode).expect("Proof creation must not fail");
-            let _ = tx.send(proof);
-        });
-
-        rx.await.expect("Should not have channel errors")
-    }
-
-    async fn submit_proofs(
-        &self,
-        proofs: Vec<Proof>,
-    ) -> anyhow::Result<Vec<<Da as DaService>::TransactionId>> {
-        let mut tx_ids = Vec::with_capacity(proofs.len());
-        for proof in proofs {
-            let tx_id = self.submit_proof(proof).await?;
-            tx_ids.push(tx_id);
-        }
-        Ok(tx_ids)
-    }
-
-    async fn submit_proof(&self, proof: Proof) -> anyhow::Result<<Da as DaService>::TransactionId> {
-        let da_data = DaData::ZKProof(proof);
-        self.da_service
-            .send_transaction(da_data)
-            .await
-            .map_err(|e| anyhow::anyhow!(e))
-    }
-}
-
-#[async_trait]
-impl<Da, Vm, Stf> ProverService2 for ParallelProverService<Da, Vm, Stf>
-where
-    Da: DaService,
-    Vm: ZkvmHost,
-    Stf: StateTransitionFunction<Vm::Guest, Da::Spec> + Send + Sync,
-    Stf::PreState: Clone + Send + Sync,
-{
-    type DaService = Da;
-
-    fn add_proof_data(&mut self, proof_data: ProofData) {
-        self.proof_queue.push(proof_data);
-    }
-
-    async fn prove_and_submit(
-        &mut self,
-    ) -> anyhow::Result<Vec<<Da as DaService>::TransactionId>> {
-        if let ProofGenMode::Skip = *self.proof_mode.lock() {
-            tracing::debug!("Skipped proving {} proofs", self.proof_queue.len(),);
-            self.proof_queue.clear();
-            return Ok(vec![]);
-        }
-
-        assert!(
-            !self.proof_queue.is_empty(),
-            "Prove should never be called before setting some proofs"
-        );
-
-        // Clear current proof data
-        let proof_queue = std::mem::take(&mut self.proof_queue);
-
-        // Prove all
-        let proofs = self.prove(proof_queue).await;
-        // Submit proofs to DA
-        self.submit_proofs(proofs).await
-    }
-}
-
-fn make_proof<Da, Vm, Stf>(
-    mut vm: Vm,
-    zk_storage: Stf::PreState,
-    proof_mode: Arc<Mutex<ProofGenMode<Da, Vm, Stf>>>,
-) -> Result<Proof, anyhow::Error>
-where
-    Da: DaService,
-    Vm: ZkvmHost,
-    Stf: StateTransitionFunction<Vm::Guest, Da::Spec> + Send + Sync,
-    Stf::PreState: Send + Sync,
-{
-    let mut proof_mode = proof_mode.lock();
-    match proof_mode.deref_mut() {
-        ProofGenMode::Skip => Ok(Vec::default()),
-        ProofGenMode::Simulate(ref mut verifier) => verifier
-            .run_sequencer_commitments_in_da_slot(vm.simulate_with_hints(), zk_storage)
-            .map(|_| Vec::default())
-            .map_err(|e| anyhow::anyhow!("Guest execution must succeed but failed with {:?}", e)),
-        ProofGenMode::Execute => vm.run(false),
-        ProofGenMode::Prove => vm.run(true),
-    }
-}
