+++ crates/sovereign-sdk/module-system/sov-modules-rollup-blueprint/src/lib.rs
@@ -54,7 +54,7 @@ pub trait RollupBlueprint: Sized + Send + Sync {
     type NativeRuntime: RuntimeTrait<Self::NativeContext, Self::DaSpec> + Default + Send + Sync;
 
     /// Prover service.
-    type ProverService: ProverService<Self::Vm, DaService = Self::DaService> + Send + Sync + 'static;
+    type ProverService: ProverService<DaService = Self::DaService> + Send + Sync + 'static;
 
     /// Creates a new instance of the blueprint.
     fn new() -> Self;
